requirements.txt
livekit>=0.17.0
livekit-agents>=0.12.0
livekit-plugins-openai>=0.8.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-cartesia>=0.3.0
livekit-plugins-silero>=0.7.0
livekit-plugins-turn-detector>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

